Made a Helper to have some helper functions grouped. Helper is used in both parseVHDL and parseExpr.
Cleaned up parseVHDL and parseExpr, deleted redunant code, made good indentation.

Deleted parseDatastruc, it was not used anymore. All the code was distributed to the other files earlier.

Made a directory for files, for example the doc about the parseArchBody.

Made the code compilable again. Fixed a small bug which involved a new variable name.